requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
Pillow==10.1.0
jinja2==3.1.2
email-validator==2.1.0
bcrypt==4.1.2
redis==5.0.1
celery==5.3.4
stripe==7.8.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
elevenlabs==0.2.26
twilio==8.11.0
requests==2.31.0

